fix(require): get rid of relative path to the word-wrap
Fix this module for npm@3 where flat directory structure is "by default". A bit of background: If the module identifier passed to require() is not a native module, and does not begin with '/', '../', or './', then Node.js starts at the parent directory of the current module, and adds /node_modules, and attempts to load the module from that location. If it is not found there, then it moves to the parent directory, and so on, until the root of the file system is reached.

This closes #4