fix(defaultValues): Added fallback for optional default values
We are using [commitizen](https://github.com/commitizen/cz-cli) and after bumping it to version 4.0.3 that bumped the version of [cz-conventional-changelog](https://github.com/commitizen/cz-conventional-changelog) to version 3.0.1 we started getting the error message .

After some investigating we could track it to the default values introduced in PR [#69](https://github.com/commitizen/cz-conventional-changelog/pull/69).

This PR makes it so the default values are optional and safeguards so if no default values are present it does not crash.