feat(engine): implement scope choices
Allow users to configure a predefined list of scopes that can be selected from, with the option to
type their own scope if it's not on the list. When passing CZ_SCOPE, we attempt to pre-select the
scope from the list of scopes, or auto-select "something else" which allows the user to confirm the
CZ_SCOPE value.