fix(deps): update semantic-release and require node v14+
This fixes a security vulnerability in the easiest possible way, by updating
`semantic-release` so it will also update its transitive dependencies. Since 14
is the LTS version as of committing, it makes sense to require v14+ in this
package, but I'm not sure how that will affect everyone. We can look for a more
"surgical" solution if that's needed.