feat(engine): add support for custom scopes option
This change is to support the custom scopes definition in the package.json's
`config.commitizen.scopes` key. The format of the key matches the format of the
`config.commitizen.types` key. Each scope has a `description` and `title` key used by inquirer to
display the scope to the user. In order to maintain backwards-compatibility, the default value for
the new package.json key (i.e., `scopes`) is an empty JSON object. When no keys are present in this
object, we will default to custom user input for the scope. This also includes the option to press
Enter to skip.

refs: #213